requirements.txt
streamlit==1.28.0
langgraph==0.0.55
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.7
langchain-ollama==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
